--- ios-sprint-9-challenge/ios-sprint-9-challenge/Views/CalorieCell.swift
@@ -1,46 +1,0 @@
-//
-//  CalorieCell.swift
-//  ios-sprint-9-challenge
-//
-//  Created by David Doswell on 9/21/18.
-//  Copyright © 2018 David Doswell. All rights reserved.
-//
-
-import UIKit
-
-class CalorieCell: UITableViewCell {
-    
-    let caloriesLabel: UILabel = {
-        let label = UILabel()
-        label.text = "Calories:"
-        label.textAlignment = .left
-        label.font = UIFont.systemFont(ofSize: 16)
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let caloriesNumber: UILabel = {
-        let label = UILabel()
-        label.textAlignment = .left
-        label.font = UIFont.systemFont(ofSize: 16)
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let timestampLabel: UILabel = {
-        let label = UILabel()
-        label.textAlignment = .right
-        label.font = UIFont.systemFont(ofSize: 16)
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: .value2, reuseIdentifier: reuseIdentifier)
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-}
